+++ MANIFEST.in
@@ -1,3 +1,3 @@
-recursive-include pandasql/data *
+recursive-include pansql/data *
 include *.txt
 include *.md
