+++ setup.py
@@ -3,14 +3,14 @@ from setuptools import find_packages
 
 setup(
     name="pansql",
-    version="0.7.4",
-    author="Greg Lamp",
-    author_email="greg@yhathq.com",
-    url="https://github.com/yhat/pandasql/",
+    version="0.0.1",
+    author="Harshad Hegde",
+    author_email="hhegde@lbl.gov",
+    url="https://github.com/hrshdhgd/pansql/",
     license="MIT",
     packages=find_packages(),
-    package_dir={"pandasql": "pandasql"},
-    package_data={"pandasql": ["data/*.csv"]},
+    package_dir={"pansql": "pansql"},
+    package_data={"pansql": ["data/*.csv"]},
     description="sqldf for pandas",
     long_description=open("README.rst").read(),
     install_requires=['numpy', 'pandas', 'sqlalchemy'],
